challenge part is running to complete

+++ src/Pages/Mytour.jsx
@@ -3,16 +3,12 @@ import React, { useContext, useEffect, useState } from 'react';
 import { ContextAPI } from '../AuthContext/ContextProvider';
 import { Link } from 'react-router-dom';
 import SingleMyTour from '../Components/SingleMyTour';
+import Swal from 'sweetalert2';
 
 function MyTourSpot() {
   const [data, setData] = useState([]);
-  const { user,loader } = useContext(ContextAPI)
+  const { user, loader } = useContext(ContextAPI);
   const Email = user?.email;
- 
-
-
-
- 
 
   useEffect(() => {
     axios
@@ -25,27 +21,46 @@ function MyTourSpot() {
       });
   }, []);
 
+  const deleteTour = id => {
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then(result => {
+      if (result.isConfirmed) {
+        axios.delete(`http://localhost:3000/tour/${id}`)
+          .then(res => {
+            const remaining = data.filter(item => item._id !== id);
+            
+            setData(remaining);
+            Swal.fire({
+              title: 'Deleted!',
+              text: 'Your file has been deleted.',
+              icon: 'success',
+            });
+        });
+        
+      }
+    });
+  };
 
-  const deleteTour = (id) => {
-    axios.delete(`http://localhost:3000/tour/${id}`)
-      .then(res => {
-        const remaining = data.filter(item => item._id !== id)
-        setData(remaining);
-    })
-  }
+  if (!data)
+    return (
+      <div className="flex justify-center items-center min-h-screen font-bold text-3xl container mx-auto">
+        <div>
+          You have no tour plan ... please add a tour plan{' '}
+          <Link className="underline text-green-500" to={'/addTourist'}>
+            {' '}
+            Add a tour{' '}
+          </Link>
+        </div>{' '}
+      </div>
+    );
 
-  if (!data) return (
-    <div className="flex justify-center items-center min-h-screen font-bold text-3xl container mx-auto">
-      <div>
-        You have no tour plan ... please add a tour plan{' '}
-        <Link className="underline text-green-500" to={'/addTourist'}>
-          {' '}
-          Add a tour{' '}
-        </Link>
-      </div>{' '}
-    </div>
-  );
- 
   return (
     <div className="lg:my-32 my-16 container mx-auto">
       <div className="text-center">
